Move acceleration structure to it's own module
