+++ src/raytracer/scene.rs
@@ -2,20 +2,10 @@ use anyhow::Result;
 use image::{GenericImageView, ImageReader};
 use std::{
     collections::HashMap,
-    iter,
-    mem::size_of,
     sync::{Arc, RwLock},
 };
 use vulkano::{
     DeviceSize,
-    acceleration_structure::{
-        AccelerationStructure, AccelerationStructureBuildGeometryInfo,
-        AccelerationStructureBuildRangeInfo, AccelerationStructureBuildType,
-        AccelerationStructureCreateInfo, AccelerationStructureGeometries,
-        AccelerationStructureGeometryInstancesData, AccelerationStructureGeometryInstancesDataType,
-        AccelerationStructureGeometryTrianglesData, AccelerationStructureInstance,
-        AccelerationStructureType, BuildAccelerationStructureFlags, BuildAccelerationStructureMode,
-    },
     buffer::{Buffer, BufferCreateInfo, BufferUsage, IndexBuffer, Subbuffer},
     command_buffer::{
         AutoCommandBufferBuilder, CommandBufferUsage, CopyBufferToImageInfo,
@@ -51,6 +41,7 @@ use vulkano::{
 
 use super::{
     Camera, MaterialPropertyData, MaterialPropertyDataEnum,
+    acceleration::AccelerationStructures,
     model::Model,
     shaders::{ShaderModules, closest_hit, ray_gen},
 };
@@ -67,10 +58,7 @@ pub struct Scene {
     memory_allocator: Arc<dyn MemoryAllocator>,
     command_buffer_allocator: Arc<dyn CommandBufferAllocator>,
 
-    // The bottom-level acceleration structure is required to be kept alive
-    // as we reference it in the top-level acceleration structure.
-    _blas_vec: Vec<Arc<AccelerationStructure>>,
-    _tlas: Arc<AccelerationStructure>,
+    acceleration_structures: AccelerationStructures,
 
     camera: Arc<RwLock<dyn Camera>>,
 }
@@ -104,77 +92,24 @@ impl Scene {
 
         let layouts = pipeline_layout.set_layouts();
 
-        // Build the bottom-level acceleration structure and then the top-level acceleration
-        // structure. Acceleration structures are used to accelerate ray tracing. The bottom-level
-        // acceleration structure contains the geometry data. The top-level acceleration structure
-        // contains the instances of the bottom-level acceleration structures. In our shader, we
-        // will trace rays against the top-level acceleration structure.
-        let vertex_buffers: Vec<_> = models
-            .iter()
-            .map(|model| {
-                model
-                    .create_blas_vertex_buffer(
-                        memory_allocator.clone(),
-                        command_buffer_allocator.clone(),
-                        queue.clone(),
-                    )
-                    .unwrap()
-            })
-            .collect();
-
-        let index_buffers: Vec<_> = models
-            .iter()
-            .map(|model| {
-                model
-                    .create_blas_index_buffer(
-                        memory_allocator.clone(),
-                        command_buffer_allocator.clone(),
-                        queue.clone(),
-                    )
-                    .unwrap()
-            })
-            .collect();
-
-        let blas_vec: Vec<_> = vertex_buffers
-            .into_iter()
-            .zip(index_buffers)
-            .map(|(vertex_buffer, index_buffer)| {
-                build_acceleration_structure_triangles(
-                    vertex_buffer,
-                    index_buffer,
-                    memory_allocator.clone(),
-                    command_buffer_allocator.clone(),
-                    device.clone(),
-                    queue.clone(),
-                )
-            })
-            .collect();
-
-        let blas_instances = blas_vec
-            .iter()
-            .map(|blas| AccelerationStructureInstance {
-                acceleration_structure_reference: blas.device_address().into(),
-                ..Default::default()
-            })
-            .collect();
-
-        // Build the top-level acceleration structure.
-        let tlas = unsafe {
-            build_top_level_acceleration_structure(
-                blas_instances,
-                memory_allocator.clone(),
-                command_buffer_allocator.clone(),
-                device.clone(),
-                queue.clone(),
-            )
-        };
-
         // For now the acceleration structure is non-changing. We can create its descriptor set
         // and clone it later during render.
+        let acceleration_structures = AccelerationStructures::new(
+            models,
+            memory_allocator.clone(),
+            command_buffer_allocator.clone(),
+            device.clone(),
+            queue.clone(),
+        )
+        .unwrap();
+
         let tlas_descriptor_set = DescriptorSet::new(
             descriptor_set_allocator.clone(),
             layouts[TLAS_LAYOUT].clone(),
-            [WriteDescriptorSet::acceleration_structure(0, tlas.clone())],
+            [WriteDescriptorSet::acceleration_structure(
+                0,
+                acceleration_structures.tlas.clone(),
+            )],
             [],
         )
         .unwrap();
@@ -299,8 +234,7 @@ impl Scene {
             pipeline,
             memory_allocator,
             command_buffer_allocator,
-            _blas_vec: blas_vec,
-            _tlas: tlas,
+            acceleration_structures,
             camera,
         }
     }
@@ -395,179 +329,6 @@ impl Scene {
 
         after_future.boxed()
     }
-}
-
-/// A helper function to build a acceleration structure and wait for its completion.
-///
-/// # Safety
-///
-/// - If you are referencing a bottom-level acceleration structure in a top-level acceleration
-///   structure, you must ensure that the bottom-level acceleration structure is kept alive.
-fn build_acceleration_structure_common(
-    geometries: AccelerationStructureGeometries,
-    primitive_count: u32,
-    ty: AccelerationStructureType,
-    memory_allocator: Arc<dyn MemoryAllocator>,
-    command_buffer_allocator: Arc<dyn CommandBufferAllocator>,
-    device: Arc<Device>,
-    queue: Arc<Queue>,
-) -> Arc<AccelerationStructure> {
-    let mut as_build_geometry_info = AccelerationStructureBuildGeometryInfo {
-        mode: BuildAccelerationStructureMode::Build,
-        flags: BuildAccelerationStructureFlags::PREFER_FAST_TRACE,
-        ..AccelerationStructureBuildGeometryInfo::new(geometries)
-    };
-
-    let as_build_sizes_info = device
-        .acceleration_structure_build_sizes(
-            AccelerationStructureBuildType::Device,
-            &as_build_geometry_info,
-            &[primitive_count],
-        )
-        .unwrap();
-
-    // We create a new scratch buffer for each acceleration structure for simplicity. You may want
-    // to reuse scratch buffers if you need to build many acceleration structures.
-    let scratch_buffer = Buffer::new_slice::<u8>(
-        memory_allocator.clone(),
-        BufferCreateInfo {
-            usage: BufferUsage::SHADER_DEVICE_ADDRESS | BufferUsage::STORAGE_BUFFER,
-            ..Default::default()
-        },
-        AllocationCreateInfo::default(),
-        as_build_sizes_info.build_scratch_size,
-    )
-    .unwrap();
-
-    let as_create_info = AccelerationStructureCreateInfo {
-        ty,
-        ..AccelerationStructureCreateInfo::new(
-            Buffer::new_slice::<u8>(
-                memory_allocator,
-                BufferCreateInfo {
-                    usage: BufferUsage::ACCELERATION_STRUCTURE_STORAGE
-                        | BufferUsage::SHADER_DEVICE_ADDRESS,
-                    ..Default::default()
-                },
-                AllocationCreateInfo::default(),
-                as_build_sizes_info.acceleration_structure_size,
-            )
-            .unwrap(),
-        )
-    };
-
-    let acceleration = unsafe { AccelerationStructure::new(device, as_create_info) }.unwrap();
-
-    as_build_geometry_info.dst_acceleration_structure = Some(acceleration.clone());
-    as_build_geometry_info.scratch_data = Some(scratch_buffer);
-
-    let as_build_range_info = AccelerationStructureBuildRangeInfo {
-        primitive_count,
-        ..Default::default()
-    };
-
-    // For simplicity, we build a single command buffer that builds the acceleration structure,
-    // then waits for its execution to complete.
-    let mut builder = AutoCommandBufferBuilder::primary(
-        command_buffer_allocator,
-        queue.queue_family_index(),
-        CommandBufferUsage::OneTimeSubmit,
-    )
-    .unwrap();
-
-    unsafe {
-        builder
-            .build_acceleration_structure(
-                as_build_geometry_info,
-                iter::once(as_build_range_info).collect(),
-            )
-            .unwrap()
-    };
-
-    builder
-        .build()
-        .unwrap()
-        .execute(queue)
-        .unwrap()
-        .then_signal_fence_and_flush()
-        .unwrap()
-        .wait(None)
-        .unwrap();
-
-    acceleration
-}
-
-fn build_acceleration_structure_triangles(
-    vertex_buffer: Subbuffer<[closest_hit::MeshVertex]>,
-    index_buffer: Subbuffer<[u32]>,
-    memory_allocator: Arc<dyn MemoryAllocator>,
-    command_buffer_allocator: Arc<dyn CommandBufferAllocator>,
-    device: Arc<Device>,
-    queue: Arc<Queue>,
-) -> Arc<AccelerationStructure> {
-    let primitive_count = (index_buffer.len() / 3) as u32;
-
-    let as_geometry_triangles_data = AccelerationStructureGeometryTrianglesData {
-        max_vertex: vertex_buffer.len() as _,
-        vertex_data: Some(vertex_buffer.into_bytes()),
-        index_data: Some(IndexBuffer::U32(index_buffer)),
-        vertex_stride: size_of::<closest_hit::MeshVertex>() as _,
-        ..AccelerationStructureGeometryTrianglesData::new(Format::R32G32B32_SFLOAT)
-    };
-
-    let geometries = AccelerationStructureGeometries::Triangles(vec![as_geometry_triangles_data]);
-
-    build_acceleration_structure_common(
-        geometries,
-        primitive_count,
-        AccelerationStructureType::BottomLevel,
-        memory_allocator,
-        command_buffer_allocator,
-        device,
-        queue,
-    )
-}
-
-unsafe fn build_top_level_acceleration_structure(
-    as_instances: Vec<AccelerationStructureInstance>,
-    allocator: Arc<dyn MemoryAllocator>,
-    command_buffer_allocator: Arc<dyn CommandBufferAllocator>,
-    device: Arc<Device>,
-    queue: Arc<Queue>,
-) -> Arc<AccelerationStructure> {
-    let primitive_count = as_instances.len() as u32;
-
-    let instance_buffer = Buffer::from_iter(
-        allocator.clone(),
-        BufferCreateInfo {
-            usage: BufferUsage::SHADER_DEVICE_ADDRESS
-                | BufferUsage::ACCELERATION_STRUCTURE_BUILD_INPUT_READ_ONLY,
-            ..Default::default()
-        },
-        AllocationCreateInfo {
-            memory_type_filter: MemoryTypeFilter::PREFER_DEVICE
-                | MemoryTypeFilter::HOST_SEQUENTIAL_WRITE,
-            ..Default::default()
-        },
-        as_instances,
-    )
-    .unwrap();
-
-    let as_geometry_instances_data = AccelerationStructureGeometryInstancesData::new(
-        AccelerationStructureGeometryInstancesDataType::Values(Some(instance_buffer)),
-    );
-
-    let geometries = AccelerationStructureGeometries::Instances(as_geometry_instances_data);
-
-    build_acceleration_structure_common(
-        geometries,
-        primitive_count,
-        AccelerationStructureType::TopLevel,
-        allocator,
-        command_buffer_allocator,
-        device,
-        queue,
-    )
 }
 
 /// Load shader modules
